Split categories to own files
