+++ app/src/main/java/org/jellyfin/androidtv/preferences/ui/preference/ButtonRemapPreference.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.preferences.ui
+package org.jellyfin.androidtv.preferences.ui.preference
 
 import android.content.Context
 import android.content.res.TypedArray
