+++ app/src/main/res/values/strings.xml
@@ -368,4 +368,7 @@
     <string name="pref_subtitle_track_button">Subtitle track button</string>
     <string name="pref_button_remapping_description">Press any button on the TV remote to assign it</string>
     <string name="lbl_Reset">Reset</string>
+    <string name="bitrate_auto">Auto</string>
+    <string name="bitrate_mbit">%s Mbit/s</string>
+    <string name="bitrate_kbit">%s Kbit/s</string>
 </resources>
